Add solution for problem `F - LCS` Link: https://atcoder.jp/contests/dp/tasks/dp_f Tags: AtCoder, *100 Ref: https://atcoder.jp/contests/dp/submissions/11400936

+++ atcoder/dp/F.cpp
@@ -21,7 +21,7 @@ int main() {
     int n = a.length();
     int m = b.length();
 
-    int dp[n+1][m+1];
+    ll dp[n+1][m+1];
 
     for(int i=0;i<=n;i++) {
     	dp[i][0] = 0;
